+++ integrations/meta_llama/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [integrations/meta_llama-v1.2.0] - 2025-10-24
+
+### 🚀 Features
+
+- `MetaLlamaChatGenerator` update tools param to ToolsType (#2437)
+
+### 📚 Documentation
+
+- Add pydoc configurations for Docusaurus (#2411)
+
+
 ## [integrations/meta_llama-v1.1.1] - 2025-09-23
 
 ### 🧹 Chores